refactor(runner): split start command into smaller modules
Extract runner logic from start.ts (377 lines) into focused modules:
- types.ts: RunnerMode, RunnerState, RunnerStatus, RunnerResources
- status.ts: writeStatusFile(), createStatusUpdater()
- setup.ts: setupEnvironment(), cleanupEnvironment()
- signals.ts: setupSignalHandlers()
- runner.ts: Runner class with lifecycle management

start.ts is now a thin CLI wrapper (~25 lines).

No behavior changes - pure refactoring.

Closes #1777
