+++ turbo/apps/runner/src/lib/runner/types.ts
@@ -1,0 +1,33 @@
+/**
+ * Runner mode for lifecycle management
+ */
+export type RunnerMode = "running" | "draining" | "stopped";
+
+/**
+ * Internal runner state shared across modules
+ */
+export interface RunnerState {
+  mode: RunnerMode;
+  activeRuns: Set<string>;
+  jobPromises: Set<Promise<void>>;
+  startedAt: Date;
+}
+
+/**
+ * Runner status for external monitoring (written to status.json)
+ * Used by deployment tools (Ansible) to track drain progress
+ */
+export interface RunnerStatus {
+  mode: RunnerMode;
+  active_runs: number;
+  active_run_ids: string[];
+  started_at: string;
+  updated_at: string;
+}
+
+/**
+ * Resources initialized during runner setup
+ */
+export interface RunnerResources {
+  proxyEnabled: boolean;
+}
